requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
django-timescaledb==0.2.13
psycopg2==2.9.9
requests==2.31.0
